float to double

+++ car.cpp
@@ -1,17 +1,17 @@
 #include <cmath>
 #include "car.hpp"
 
-car::car(float wb, float rh, float hx, float hy, float hdirection, float halpha) : position(hx, hy,hdirection){
+car::car(double wb, double rh, double hx, double hy, double hdirection, double halpha) : position(hx, hy,hdirection){
     wheelbase = wb;
     rear2hitch = rh;
     alpha = halpha;
 }
 
-float car::get_r2h(){
+double car::get_r2h(){
     return rear2hitch;
 }
 
-float car::get_wb(){
+double car::get_wb(){
     return wheelbase;
 }
 
@@ -29,7 +29,7 @@ point car::get_hitch(){ // needs testing
     return ret_val;
 }
 
-void car::calc_curve(float lenght, float alpha_steer, float &x, float &y, float &angle) // works for me
+void car::calc_curve(double lenght, double alpha_steer, double &x, double &y, double &angle) // works for me
 {
     if(alpha_steer == 0.0f){
         x=lenght;
@@ -37,15 +37,15 @@ void car::calc_curve(float lenght, float alpha_steer, float &x, float &y, float 
         angle = 0;
     }
     else{
-        float r_bw = wheelbase / tan(alpha_steer);
+        double r_bw = wheelbase / tan(alpha_steer);
         angle = lenght/r_bw;
         x = sin(angle) * r_bw;
         y = (1 - cos(angle)) * r_bw;// * SIGN(alpha_steer);
     }
 }
 
-void car::move(float move_lenght){ // Works
-    float x,y,a;
+void car::move(double move_lenght){ // Works
+    double x,y,a;
     calc_curve(move_lenght,alpha,x,y,a); //Works
     move_straight(x,y);
     direction+=a;
